+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `VerticalProgress` : Add `withoutBorder` prop.
+
 ## [3.22.1] - 2021-07-01
 
 Fix: